requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0